use the mapping catalog reader when backing up
This is useful if a backup is taken both from Unix and from Windows with a shared filer mounted at different places, for example. On Windows the filer may show up as F:\\, whereas on Unix it is mounted as /mnt/filer. By providing the proper mappings:

mappings:
  /mnt/filer: F:\
  "F:\": /mnt/filer

Gletscher will be able to very efficiently perform backups even if we alternate from Unix to Windows on a daily basis.